Add files via upload
A currency converter is basically a program designed to convert the value of one currency into another based on the market rate for that currency. The trick is where those values are generated. The values are usually based on the current market values or bank exchange rates, and this can make all the difference.