Add files via upload
A stopwatch is a handheld timepiece designed to gauge the span of time which elapses between its activation and deactivation. In manual stopwatch time, the watch is started and stopped by a user pressing a button. However, in fully automatic timing, both starting and stopping are stimulated automatically, by sensors. Mechanical stopwatches are powered by a mainspring that should be wound up by turning the knurled knob located at the top of the stopwatch.